Adds an experimental (default disabled) "server pauser"
This essentially lengthens tick times drastically when no player is
online, effectively pausing the game.  This will not respond well to
things that monitor times across multiple ticks, and is mostly intended
for modpacks where that doesn't cause trouble, where the players want to
be able to keep a server running, but not have time pass while noone is
online.

Example usage: Player has chunkloaders in base, so that machines keep
running while Player is out mining/exploring/etc.  Player logs off for
the night, server tps slows down by a factor of 100, when Player logs
back on, most of the fuel in the base is still there.

Use-case where this is useless: Player's base will last 2 years without
re-fueling, Player wants to leave stuff processing, Player logs out for
8 hours, when Player logs back in, only 5 minutes have passed on server.